Move utility classes to their own package.
