+++ src/ispd/gui/CreateSchedulerDialog.java
@@ -4,6 +4,7 @@ import ispd.alocacaoVM.ManipularArquivosAlloc;
 import ispd.arquivo.interpretador.gerador.InterpretadorGerador;
 import ispd.escalonador.ManipularArquivos;
 import ispd.escalonadorCloud.ManipularArquivosCloud;
+import ispd.gui.utils.ButtonBuilder;
 import ispd.utils.ValidaValores;
 
 import javax.swing.AbstractListModel;
