+++ src/ispd/gui/utils/ButtonBuilder.java
@@ -1,4 +1,4 @@
-package ispd.gui;
+package ispd.gui.utils;
 
 import javax.swing.Icon;
 import javax.swing.JButton;
@@ -19,12 +19,13 @@ public class ButtonBuilder {
         this.button = new JButton(text);
         this.button.addActionListener(onClick);
     }
+
     private ButtonBuilder(final Icon icon, final ActionListener onClick){
         this.button = new JButton(icon);
         this.button.addActionListener(onClick);
     }
 
-    static JButton basicButton(
+    public static JButton basicButton(
             final String text, final ActionListener onClick) {
         return ButtonBuilder.aButton(text, onClick).build();
     }
@@ -33,32 +34,32 @@ public class ButtonBuilder {
         return this.button;
     }
 
-    static ButtonBuilder aButton(
+    public static ButtonBuilder aButton(
             final String text, final ActionListener onClick) {
         return new ButtonBuilder(text, onClick);
     }
 
-    static ButtonBuilder aButton(
+    public static ButtonBuilder aButton(
             final Icon icon, final ActionListener onClick) {
         return new ButtonBuilder(icon, onClick);
     }
 
-    ButtonBuilder withToolTip(final String theText) {
+    public ButtonBuilder withToolTip(final String theText) {
         this.button.setToolTipText(theText);
         return this;
     }
 
-    ButtonBuilder withIcon(final Icon theIcon) {
+    public ButtonBuilder withIcon(final Icon theIcon) {
         this.button.setIcon(theIcon);
         return this;
     }
 
-    ButtonBuilder nonFocusable() {
+    public ButtonBuilder nonFocusable() {
         this.button.setFocusable(false);
         return this;
     }
 
-    ButtonBuilder withSize(final Dimension theSize) {
+    public ButtonBuilder withSize(final Dimension theSize) {
         // TODO: Find out if these three together are equivalent to setSize()
         this.button.setMaximumSize(theSize);
         this.button.setMinimumSize(theSize);
@@ -66,23 +67,23 @@ public class ButtonBuilder {
         return this;
     }
 
-    ButtonBuilder withPreferredSize(final Dimension theSize) {
+    public ButtonBuilder withPreferredSize(final Dimension theSize) {
         this.button.setPreferredSize(theSize);
         return this;
     }
 
-    ButtonBuilder withCenterBottomTextPosition() {
+    public ButtonBuilder withCenterBottomTextPosition() {
         this.button.setHorizontalTextPosition(SwingConstants.CENTER);
         this.button.setVerticalTextPosition(SwingConstants.BOTTOM);
         return this;
     }
 
-    ButtonBuilder disabled() {
+    public ButtonBuilder disabled() {
         this.button.setEnabled(false);
         return this;
     }
 
-    ButtonBuilder withActionCommand(final String theText) {
+    public ButtonBuilder withActionCommand(final String theText) {
         this.button.setActionCommand(theText);
         return this;
     }
